requirements.txt
# Core Flask dependencies
Flask>=2.3,<4
Werkzeug>=2.3,<4

# Authentication and security
Authlib>=1.3,<2
cryptography>=3.4,<4
Flask-WTF>=1.2,<2
Flask-Limiter>=3.13,<4

# Environment and configuration
python-dotenv>=1.0,<2

# HTTP requests
requests>=2.31,<3
urllib3>=2.0,<3

# Database (SQLite is built-in, but we might want to add migrations later)
# alembic>=1.12,<2  # Uncomment when adding database migrations

# Development and testing (optional)
# pytest>=7.0,<8
# pytest-flask>=1.2,<2
# black>=23.0,<24  # Code formatting
# flake8>=6.0,<7   # Linting

# Production (optional)
# gunicorn>=21.0,<22  # WSGI server
# gevent>=23.0,<24     # Async support
